requirements.txt
Flask
google-generativeai
gTTS
Pillow
pygame
pyautogui
pyttsx3
requests
SpeechRecognition
tk
wikipedia
Werkzeug
python-dateutil
shutil